add hella logging
This commit pulls in the winston package for more comprehensive,
structured logging, replacing the existing `console.log` invocations on
connect, disconnect and ifttt request error.

This also adds new messages on startup just echoing the configuration,
and before sending an ifttt notifaction with the channel/sender/message
text.

If the log level is adjustd to `debug` using the `LOG_LEVEL` environment
variable, we also log every incoming message and our attempts to match
it against the monitored terms/channels.